# restbackend/src/main/resources/openapispec/openapi.yaml
{"openapi":"3.0.1","info":{"title":"Backend API","description":"TEST BACKEND API","version":"2.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/api/users":{"get":{"tags":["user-api"],"operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"post":{"tags":["user-api"],"operationId":"addUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/uploadFile":{"post":{"tags":["file-api"],"operationId":"uploadFile","requestBody":{"content":{"application/json":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FileUploadResponse"}}}}}}},"/api/auth/signup":{"post":{"tags":["auth-api"],"operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/auth/signin":{"post":{"tags":["auth-api"],"operationId":"authenticateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JwtResponse"}}}}}}},"/api/users/{id}":{"get":{"tags":["user-api"],"operationId":"getuserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"minimum":1,"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}},"delete":{"tags":["user-api"],"operationId":"deleteEmployee","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/users/email":{"get":{"tags":["user-api"],"operationId":"getuserByEmail","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/no_auth/users":{"get":{"tags":["no-auth-api"],"operationId":"getAllUsers_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}}}}},"/api/no_auth/users/{id}":{"get":{"tags":["no-auth-api"],"operationId":"getuserById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"minimum":1,"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/no_auth/users/email":{"get":{"tags":["no-auth-api"],"operationId":"getuserByEmail_1","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/downloadFile/{fileCode}":{"get":{"tags":["file-api"],"operationId":"downloadFile","parameters":[{"name":"fileCode","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"AdressRequest":{"type":"object","properties":{"description":{"type":"string"},"country":{"type":"string"},"city":{"type":"string"},"street":{"type":"string"},"houseNumber":{"type":"integer","format":"int32"},"zipCode":{"type":"string"}}},"UserRequest":{"type":"object","properties":{"name":{"type":"string"},"username":{"type":"string"},"email":{"type":"string"},"roles":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"adresses":{"type":"array","items":{"$ref":"#/components/schemas/AdressRequest"}}}},"UserResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"username":{"type":"string"},"email":{"type":"string"},"roles":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"adresses":{"type":"array","items":{"$ref":"#/components/schemas/AdressRequest"}}}},"FileUploadResponse":{"type":"object","properties":{"fileName":{"type":"string"},"size":{"type":"integer","format":"int64"},"downloadUri":{"type":"string"}}},"SignUpRequest":{"required":["email","name","password","username"],"type":"object","properties":{"name":{"maxLength":50,"minLength":3,"type":"string"},"username":{"maxLength":50,"minLength":3,"type":"string"},"email":{"maxLength":60,"minLength":0,"type":"string"},"password":{"maxLength":40,"minLength":6,"type":"string"},"adresses":{"type":"array","items":{"$ref":"#/components/schemas/AdressRequest"}}}},"LoginRequest":{"required":["password","username"],"type":"object","properties":{"username":{"maxLength":60,"minLength":3,"type":"string"},"password":{"maxLength":40,"minLength":6,"type":"string"}}},"JwtResponse":{"type":"object","properties":{"token":{"type":"string"},"type":{"type":"string"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}